// frontend/src/api/types/CreateCategoryRequest.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type CreateCategoryRequest = { name: string, description: string, 
/**
 * Value names in order.
 */
values: Array<string>, required_everywhere: boolean, required_topic_ids: Array<string>, };